+++ app/Http/Controllers/user/CheckoutController.php
@@ -3,13 +3,15 @@
 namespace App\Http\Controllers\user;
 
 use App\Interfaces\SessionKeyInterface;
-use App\Models\Order;
-use App\Models\Shipping;
+use App\Interfaces\StatusInterface;
+use App\Models\ErrorLog;
 use App\Services\CheckoutService;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Response;
 
-class CheckoutController extends Controller implements SessionKeyInterface
+class CheckoutController extends Controller implements SessionKeyInterface, StatusInterface
 {
     private $checkoutService;
 
@@ -25,29 +27,88 @@ class CheckoutController extends Controller implements SessionKeyInterface
      * Summary of index
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
-    public function index()
-    {
-        logger('Refresh');
-        $shippings = $this->checkoutService->getCheckoutCredentials();
-        $order = $this->checkoutService->createOrder();
+    // public function index()
+    // {
+    //     $shippings = $this->checkoutService->getCheckoutCredentials();
+    //     $order = $this->checkoutService->createOrderFromCart();
 
-        session([self::SESSION_CHECKOUT => true]);
+    //     return view('checkout.index', compact('shippings', 'order'));
+    // }
+
+    /**
+     * Summary of createOrderFromCart
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\RedirectResponse
+     */
+    public function createOrderFromCart()
+    {
+        try {
+            DB::beginTransaction();
+
+            $shippings = $this->checkoutService->getCheckoutCredentials();
+            $order = $this->checkoutService->createOrderFromCart();
+
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollBack();
+
+            $errorLog = new ErrorLog();
+            $errorLog->error = $e->getMessage();
+            $errorLog->save();
+
+            $response = [
+                'status' => self::STATUS_ERROR,
+                'message' => __('message.invalid'),
+            ];
+
+            return back()->withInput()->with($response);
+        }
 
         return view('checkout.index', compact('shippings', 'order'));
     }
 
+    /**
+     * Summary of updateShipping
+     * @param string $id
+     * @param string $shipping
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\RedirectResponse
+     */
     public function updateShipping($id, $shipping)
     {
         if (!request()->ajax()) abort(404);
 
-        $this->checkoutService->updateShipping($id, $shipping);
+        try {
+            DB::beginTransaction();
+
+            $this->checkoutService->updateShipping($id, $shipping);
+
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollBack();
+
+            $errorLog = new ErrorLog();
+            $errorLog->error = $e->getMessage();
+            $errorLog->save();
+
+            $response = [
+                'status' => self::STATUS_ERROR,
+                'message' => __('message.invalid'),
+            ];
+
+            return back()->withInput()->with($response);
+        }
+
+        return response()->json([
+            'status' => self::STATUS_SUCCESS,
+            'message' => 'Update Shipping Success',
+            'data' => [],
+        ], Response::HTTP_OK);
     }
 
-    public function pay()
-    {
-        if (!session(self::SESSION_CHECKOUT)) abort(404);
-        session()->forget(self::SESSION_CHECKOUT);
+    // public function pay()
+    // {
+    //     if (!session(self::SESSION_CHECKOUT)) abort(404);
+    //     session()->forget(self::SESSION_CHECKOUT);
 
         
-    }
+    // }
 }
