+++ resources/views/checkout/component/__shipping.blade.php
@@ -7,7 +7,7 @@
         class="flex items-center justify-between w-full px-4 py-6 ms-2 text-sm font-medium text-gray-900">
         <div class="flex items-center gap-4 w-full">
             @include("checkout.component.icon.$shipping->icon")
-            <div class="md:flex w-full justify-between">
+            <div class="md:flex w-full justify-between items-center">
                 <div>
                     <h2 class="text-lg font-bold">{{ $shipping->name }}</h2>
                     <p class="text-xs text-font_secondary">({{ $shipping->description }})</p>
